Adapt tests to new rails based submenu of notification center
