+++ config/locales/en.yml
@@ -1903,6 +1903,13 @@ en:
         unknown_storage: "Unknown storage"
 
   notifications:
+    reasons:
+      assigned: "Assignee"
+      date_alert: "Date alert"
+      mentioned: "Mentioned"
+      responsible: "Accountable"
+      shared: "Shared"
+      watched: "Watcher"
     menu:
       by_project: "Unread by project"
       by_reason: "Reason"
@@ -2593,7 +2600,6 @@ en:
         shared: "Shared"
         subscribed: "all"
         prefix: "Received because of the notification setting: %{reason}"
-        date_alert: "Date alert"
         date_alert_start_date: "Date alert"
         date_alert_due_date: "Date alert"
       see_all: "See all"
