+++ spec/support/components/common/submenu.rb
@@ -32,6 +32,10 @@ module Components
     include Capybara::RSpecMatchers
     include RSpec::Matchers
 
+    def expect_open
+      expect(page).to have_css('[data-test-selector="op-submenu"]')
+    end
+
     def expect_item(name, selected: false, favored: nil, visible: true)
       within "#main-menu" do
         selected_specifier = selected ? ".selected" : ":not(.selected)"
@@ -56,6 +60,18 @@ module Components
       end
     end
 
+    def expect_item_with_count(item, count)
+      within page.find_test_selector("op-sidemenu--item-action", text: item) do
+        expect_count count
+      end
+    end
+
+    def expect_item_with_no_count(item)
+      within page.find_test_selector("op-sidemenu--item-action", text: item) do
+        expect_no_count
+      end
+    end
+
     def click_item(name)
       within "#main-menu" do
         click_on text: name
@@ -74,10 +90,22 @@ module Components
       end
     end
 
+    def finished_loading
+      wait_for_network_idle if using_cuprite?
+    end
+
     def expect_no_results_text
       within "#main-menu" do
         expect(page).to have_test_selector("op-sidebar--search-no-results", text: "No items found")
       end
+    end
+
+    def expect_count(count)
+      expect(page).to have_test_selector("op-sidemenu--item-count", text: count)
+    end
+
+    def expect_no_count
+      expect(page).not_to have_test_selector("op-sidemenu--item-count")
     end
   end
 end
